refactor: common/configuration -> internal/configuration
