+++ backend/controller/admin/local_client_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/alecthomas/assert/v2"
 	"github.com/alecthomas/types/optional"
 
-	cf "github.com/TBD54566975/ftl/common/configuration"
+	cf "github.com/TBD54566975/ftl/internal/configuration"
 	in "github.com/TBD54566975/ftl/internal/integration"
 	"github.com/TBD54566975/ftl/internal/log"
 )
