+++ internal/configuration/db_secret_resolver.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/alecthomas/types/optional"
 
-	"github.com/TBD54566975/ftl/common/configuration/dal"
+	"github.com/TBD54566975/ftl/internal/configuration/dal"
 )
 
 // DBSecretResolver loads values a project's secrets from the given database.
