+++ internal/configuration/keychain_provider.go
