+++ internal/configuration/manager.go
