+++ internal/configuration/manager_test.go
