+++ internal/configuration/manual_sync_utils.go
