+++ internal/configuration/obfuscator.go
