+++ internal/configuration/obfuscator_test.go
