+++ internal/configuration/projectconfig_resolver.go
