+++ internal/configuration/projectconfig_resolver_test.go
