+++ internal/configuration/sql/queries.sql
