+++ internal/configuration/sql/queries.sql.go
