+++ internal/configuration/testdata/ftl-project.toml
